backend/src/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djoser==2.1.0
flake8==4.0.1
isort==5.10.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.0.4
